backend/requirements.txt
# Jukeyman Autonomous Media Station (JAMS) - Backend Dependencies

# Core Framework
fastapi[all]==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis & Caching
redis==5.0.4
hiredis==2.3.2

# Celery (Task Queue)
celery[redis]==5.4.0
flower==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic[email]==2.7.1

# AWS/S3 (Cloudflare R2)
boto3==1.34.100
botocore==1.34.100

# HTTP Clients
httpx==0.27.0
requests==2.32.2
aiohttp==3.9.5

# Image Processing
Pillow==10.3.0
opencv-python==4.9.0.80

# AI/ML Libraries
torch==2.3.0
torchvision==0.18.0
torchaudio==2.3.0
transformers==4.40.2
diffusers==0.27.2
accelerate==0.30.0

# ComfyUI Client
websocket-client==1.8.0

# Payment Processing
stripe==9.6.0

# Social Media APIs
tweepy==4.14.0
praw==7.7.1

# Monitoring
prometheus-client==0.20.0
sentry-sdk[fastapi]==2.3.1

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1
ujson==5.10.0
orjson==3.10.3
python-slugify==8.0.4
validators==0.28.1

# Rate Limiting
slowapi==0.1.9
fastapi-limiter==0.1.6

# CORS
fastapi-cors==0.0.6

# Email
fastapi-mail==1.4.1

# WebSockets
python-socketio==5.11.2

# OpenAI/LLM
openai==1.30.1
anthropic==0.25.8

# Code Quality
black==24.4.2
flake8==7.0.0
mypy==1.10.0
pytest==8.2.1
pytest-asyncio==0.23.6

